backend/requirements.txt
fastapi==0.128.0
uvicorn==0.39.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

faiss-cpu==1.13.0

numpy==1.26.4
joblib==1.3.2
threadpoolctl==3.2.0

transformers==4.57.3
tokenizers==0.22.1
huggingface-hub==0.36.0
safetensors==0.7.0

pdfplumber==0.11.8
pypdfium2==5.2.0

python-dotenv==1.2.1
requests==2.32.5
pyyaml==6.0.3
regex==2024.5.15

annotated-types==0.7.0
anyio==4.12.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.1
filelock==3.19.1
fsspec==2024.10.0
h11==0.16.0
idna==3.11
jinja2==3.1.6
markupsafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
packaging==25.0
pydantic==2.12.5
pydantic-core==2.41.5
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.2
tqdm==4.67.1

python-multipart==0.0.9
groq==0.9.0
httpx==0.27.2

# Going to avoid using this dependency because it is quite large, 300mb...
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.6.0+cpu
sentence-transformers==3.0.1